+++ projects/egeo/src/lib/st-vertical-icon-tabs/st-vertical-icon-tabs.component.ts
@@ -1,13 +1,14 @@
 /*
- * © 2017 Stratio Big Data Inc., Sucursal en España. All rights reserved.
+ * © 2017 Stratio Big Data Inc., Sucursal en España.
  *
- * This software – including all its source code – contains proprietary
- * information of Stratio Big Data Inc., Sucursal en España and
- * may not be revealed, sold, transferred, modified, distributed or
- * otherwise made available, licensed or sublicensed to third parties;
- * nor reverse engineered, disassembled or decompiled, without express
- * written authorization from Stratio Big Data Inc., Sucursal en España.
+ * This software is licensed under the Apache License, Version 2.0.
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the terms of the License for more details.
+ *
+ * SPDX-License-Identifier: Apache-2.0.
  */
+
 import { ChangeDetectionStrategy, Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 
 import { StIconTab } from './st-icon-tabs.model';